Spec-compliant implementation of Eth1 monitoring; Eth1-enabled local sim
BEWARE! This commit will trigger a stack overflow during local sim

+++ tests/simulation/start.sh
@@ -32,6 +32,86 @@ else
   EXE_SUFFIX=""
 fi
 
+# to allow overriding the program names
+MULTITAIL="${MULTITAIL:-multitail}"
+TMUX="${TMUX:-tmux}"
+GANACHE="${GANACHE:-ganache-cli}"
+PROMETHEUS="${PROMETHEUS:-prometheus}"
+TMUX_SESSION_NAME="${TMUX_SESSION_NAME:-nbc-sim}"
+
+WAIT_GENESIS="${WAIT_GENESIS:-no}"
+
+# Using tmux or multitail is an opt-in
+USE_MULTITAIL="${USE_MULTITAIL:-no}"
+type "$MULTITAIL" &>/dev/null || { echo "${MULTITAIL}" is missing; USE_MULTITAIL="no"; }
+
+USE_TMUX="${USE_TMUX:-no}"
+type "$TMUX" &>/dev/null || { echo "${TMUX}" is missing; USE_TMUX="no"; }
+
+USE_GANACHE="${USE_GANACHE:-no}"
+type "$GANACHE" &>/dev/null || { echo $GANACHE is missing; USE_GANACHE="no"; }
+
+USE_PROMETHEUS="${LAUNCH_PROMETHEUS:-no}"
+type "$PROMETHEUS" &>/dev/null || { echo $PROMETHEUS is missing; USE_PROMETHEUS="no"; }
+
+# Prometheus config (continued inside the loop)
+mkdir -p "${METRICS_DIR}"
+cat > "${METRICS_DIR}/prometheus.yml" <<EOF
+global:
+  scrape_interval: 1s
+
+scrape_configs:
+  - job_name: "nimbus"
+    static_configs:
+EOF
+
+for i in $(seq $MASTER_NODE -1 $TOTAL_USER_NODES); do
+  # Prometheus config
+  cat >> "${METRICS_DIR}/prometheus.yml" <<EOF
+      - targets: ['127.0.0.1:$(( BASE_METRICS_PORT + i ))']
+        labels:
+          node: '$i'
+EOF
+done
+
+COMMANDS=()
+
+if [[ "$USE_TMUX" != "no" ]]; then
+  $TMUX new-session -s "${TMUX_SESSION_NAME}" -d
+
+  # maybe these should be moved to a user config file
+  $TMUX set-option -t "${TMUX_SESSION_NAME}" history-limit 999999
+  $TMUX set-option -t "${TMUX_SESSION_NAME}" remain-on-exit on
+  $TMUX set -t "${TMUX_SESSION_NAME}" mouse on
+
+  # We create a new window, so the above settings can take place
+  $TMUX new-window -d -t "${TMUX_SESSION_NAME}" -n "sim"
+
+  trap 'tmux kill-session -t "${TMUX_SESSION_NAME}"' SIGINT EXIT
+fi
+
+if [[ "$USE_GANACHE" != "no" ]]; then
+  if [[ "$USE_TMUX" != "no" ]]; then
+    $TMUX new-window -d -t $TMUX_SESSION_NAME -n "$GANACHE" "$GANACHE"
+  elif [[ "$USE_MULTITAIL" != "no" ]]; then
+    COMMANDS+=( " -cT ansi -t '$GANACHE'" )
+  else
+    $GANACHE &
+  fi
+fi
+
+if [[ "$USE_PROMETHEUS" != "no" ]]; then
+  if [[ "$USE_TMUX" != "no" ]]; then
+    $TMUX new-window -d -t $TMUX_SESSION_NAME -n "$PROMETHEUS" "cd '$METRICS_DIR' && $PROMETHEUS"
+  else
+    echo "$PROMETHEUS can be used currently only with USE_TMUX=1"
+  fi
+fi
+
+if [[ "$USE_TMUX" != "no" ]]; then
+  $TMUX select-window -t "${TMUX_SESSION_NAME}:sim"
+fi
+
 build_beacon_node () {
   OUTPUT_BIN=$1; shift
   PARAMS="$CUSTOM_NIMFLAGS $DEFS $@"
@@ -45,29 +125,48 @@ if [ ! -f "${LAST_VALIDATOR}" ]; then
   echo Building "${DEPLOY_DEPOSIT_CONTRACT_BIN}"
   $MAKE NIMFLAGS="-o:\"$DEPLOY_DEPOSIT_CONTRACT_BIN\" $CUSTOM_NIMFLAGS $DEFS" deposit_contract
 
-  if [ "$DEPOSIT_WEB3_URL_ARG" != "" ]; then
-    DEPOSIT_CONTRACT_ADDRESS=$($DEPLOY_DEPOSIT_CONTRACT_BIN deploy $DEPOSIT_WEB3_URL_ARG)
+  if [ "$WEB3_ARG" != "" ]; then
+    echo Deploying the validator deposit contract...
+    DEPOSIT_CONTRACT_ADDRESS=$($DEPLOY_DEPOSIT_CONTRACT_BIN deploy $WEB3_ARG)
+    echo Contract deployed at $DEPOSIT_CONTRACT_ADDRESS
     export DEPOSIT_CONTRACT_ADDRESS
+  fi
+
+  DELAY_ARGS=""
+
+  # Uncomment this line to slow down the initial deposits.
+  # This will spread them across multiple blocks which is
+  # a more realistic scenario.
+  DELAY_ARGS="--min-delay=1 --max-delay=5"
+
+  MAKE_DEPOSITS_WEB3_ARG=$WEB3_ARG
+  if [[ "$WAIT_GENESIS" == "no" ]]; then
+    MAKE_DEPOSITS_WEB3_ARG=""
   fi
 
   $BEACON_NODE_BIN makeDeposits \
     --quickstart-deposits="${NUM_VALIDATORS}" \
     --deposits-dir="$VALIDATORS_DIR" \
-    $DEPOSIT_WEB3_URL_ARG \
+    $MAKE_DEPOSITS_WEB3_ARG $DELAY_ARGS \
     --deposit-contract="${DEPOSIT_CONTRACT_ADDRESS}"
+
+  echo "All deposits prepared"
 fi
 
 if [ ! -f "${SNAPSHOT_FILE}" ]; then
-  $BEACON_NODE_BIN \
-    --data-dir="${SIMULATION_DIR}/node-$MASTER_NODE" \
-    createTestnet \
-    --validators-dir="${VALIDATORS_DIR}" \
-    --total-validators="${NUM_VALIDATORS}" \
-    --output-genesis="${SNAPSHOT_FILE}" \
-    --output-bootstrap-file="${NETWORK_BOOTSTRAP_FILE}" \
-    --bootstrap-address=127.0.0.1 \
-    --bootstrap-port=$(( BASE_P2P_PORT + MASTER_NODE )) \
-    --genesis-offset=5 # Delay in seconds
+  if [[ "${WAIT_GENESIS}" == "no" ]]; then
+    echo Creating testnet genesis...
+    $BEACON_NODE_BIN \
+      --data-dir="${SIMULATION_DIR}/node-$MASTER_NODE" \
+      createTestnet \
+      --validators-dir="${VALIDATORS_DIR}" \
+      --total-validators="${NUM_VALIDATORS}" \
+      --output-genesis="${SNAPSHOT_FILE}" \
+      --output-bootstrap-file="${NETWORK_BOOTSTRAP_FILE}" \
+      --bootstrap-address=127.0.0.1 \
+      --bootstrap-port=$(( BASE_P2P_PORT + MASTER_NODE )) \
+      --genesis-offset=5 # Delay in seconds
+  fi
 fi
 
 rm -f beacon_node.log
@@ -77,29 +176,6 @@ if [ -f "${MASTER_NODE_ADDRESS_FILE}" ]; then
   rm "${MASTER_NODE_ADDRESS_FILE}"
 fi
 
-# to allow overriding the program names
-MULTITAIL="${MULTITAIL:-multitail}"
-TMUX="${TMUX:-tmux}"
-TMUX_SESSION_NAME="${TMUX_SESSION_NAME:-nbc-network-sim}"
-
-# Using tmux or multitail is an opt-in
-USE_MULTITAIL="${USE_MULTITAIL:-no}"
-type "$MULTITAIL" &>/dev/null || { echo "${MULTITAIL}" is missing; USE_MULTITAIL="no"; }
-
-USE_TMUX="${USE_TMUX:-no}"
-type "$TMUX" &>/dev/null || { echo "${TMUX}" is missing; USE_TMUX="no"; }
-
-# Prometheus config (continued inside the loop)
-mkdir -p "${METRICS_DIR}"
-cat > "${METRICS_DIR}/prometheus.yml" <<EOF
-global:
-  scrape_interval: 1s
-
-scrape_configs:
-  - job_name: "nimbus"
-    static_configs:
-EOF
-
 PROCESS_DASHBOARD_BIN="build/process_dashboard${EXE_SUFFIX}"
 
 if [[ ! -f "$PROCESS_DASHBOARD_BIN" ]]; then
@@ -107,6 +183,7 @@ if [[ ! -f "$PROCESS_DASHBOARD_BIN" ]]; then
 fi
 
 # use the exported Grafana dashboard for a single node to create one for all nodes
+echo Creating grafana dashboards...
 "${PROCESS_DASHBOARD_BIN}" \
   --nodes=${TOTAL_NODES} \
   --in="${SIM_ROOT}/beacon-chain-sim-node0-Grafana-dashboard.json" \
@@ -115,26 +192,20 @@ fi
 # Kill child processes on Ctrl-C by sending SIGTERM to the whole process group,
 # passing the negative PID of this shell instance to the "kill" command.
 # Trap and ignore SIGTERM, so we don't kill this process along with its children.
-if [ "$USE_MULTITAIL" = "no" ]; then
-  trap '' SIGTERM
+if [[ "$USE_MULTITAIL" == "no" && "$USE_TMUX" == "no" ]]; then
   trap 'pkill -P $$ beacon_node' SIGINT EXIT
 fi
 
-COMMANDS=()
-
-if [[ "$USE_TMUX" != "no" ]]; then
-  $TMUX new-session -s "${TMUX_SESSION_NAME}" -d
-
-  # maybe these should be moved to a user config file
-  $TMUX set-option -t "${TMUX_SESSION_NAME}" history-limit 999999
-  $TMUX set-option -t "${TMUX_SESSION_NAME}" remain-on-exit on
-  $TMUX set -t "${TMUX_SESSION_NAME}" mouse on
-fi
+LAST_WAITING_NODE=0
 
 for i in $(seq $MASTER_NODE -1 $TOTAL_USER_NODES); do
   if [[ "$i" != "$MASTER_NODE" && "$USE_MULTITAIL" == "no" ]]; then
     # Wait for the master node to write out its address file
     while [ ! -f "${MASTER_NODE_ADDRESS_FILE}" ]; do
+      if (( LAST_WAITING_NODE != i )); then
+        echo Waiting for $MASTER_NODE_ADDRESS_FILE to appear...
+        LAST_WAITING_NODE=i
+      fi
       sleep 0.1
     done
   fi
@@ -142,6 +213,8 @@ for i in $(seq $MASTER_NODE -1 $TOTAL_USER_NODES); do
   CMD="${SIM_ROOT}/run_node.sh ${i} --verify-finalization"
 
   if [[ "$USE_TMUX" != "no" ]]; then
+    echo "Starting node $i..."
+    echo $TMUX split-window -t "${TMUX_SESSION_NAME}" "$CMD"
     $TMUX split-window -t "${TMUX_SESSION_NAME}" "$CMD"
     $TMUX select-layout -t "${TMUX_SESSION_NAME}" tiled
   elif [[ "$USE_MULTITAIL" != "no" ]]; then
@@ -155,16 +228,13 @@ for i in $(seq $MASTER_NODE -1 $TOTAL_USER_NODES); do
   else
     eval "${CMD}" &
   fi
-
-  # Prometheus config
-  cat >> "${METRICS_DIR}/prometheus.yml" <<EOF
-      - targets: ['127.0.0.1:$(( BASE_METRICS_PORT + i ))']
-        labels:
-          node: '$i'
-EOF
 done
 
 if [[ "$USE_TMUX" != "no" ]]; then
+  # kill the console window in the pane where the simulation is running
+  $TMUX kill-pane -t $TMUX_SESSION_NAME:sim.0
+  # kill the original console window
+  # (this one doesn't have the right history-limit)
   $TMUX kill-pane -t $TMUX_SESSION_NAME:0.0
   $TMUX select-layout -t "${TMUX_SESSION_NAME}" tiled
   $TMUX attach-session -t "${TMUX_SESSION_NAME}" -d
